pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "doco-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic==0.91.0",
    "pydantic==2.12.5",
    "fastapi==0.135.3",
    "uvicorn[standard]==0.44.0",
    "httpx==0.28.1",
    "python-dotenv==1.2.2",
    "python-multipart>=0.0.9",
    "slack-sdk==3.41.0",
    "boto3>=1.34",
]

[project.optional-dependencies]
test = ["pytest==9.0.3", "moto[dynamodb,sqs]>=5.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
